feat: add consolidated geo module and tests
